requirements.txt
monai~=1.3.0
numpy~=1.26.4
torch~=2.2.1
antspyx~=0.4.2
